+++ src/break-sentence.c
@@ -76,7 +76,8 @@ static inline bool mjb_peek_lower_sentence(const char *buffer, size_t byte_lengt
 // See: https://unicode.org/reports/tr29/
 MJB_EXPORT mjb_break_type mjb_next_sentence_break(const char *buffer, size_t byte_length,
     mjb_encoding encoding, mjb_next_sentence_state *state) {
-    if(buffer == NULL || state == NULL || byte_length == 0) {
+    if(buffer == NULL || state == NULL || byte_length == 0 ||
+        byte_length == MJB_NUL_TERMINATED) {
         return MJB_BT_NOT_SET;
     }